minor: improve message display for tools/openstack-cleanup.sh

+++ tools/openstack-cleanup.sh
@@ -15,18 +15,22 @@ if openstack ${OS_ARGS} endpoint list &> /dev/null; then
     exit 1
 fi
 
-echo -e "\U0001F5D1 Start cleanup for tag: ${CAPO_TAG} at $(date)"
+echo -e "\U0001F5D1 Start cleanup for tag '${CAPO_TAG}' at $(date)"
 
 SERVERS="$(openstack ${OS_ARGS} server list --tags ${CAPO_TAG} -f value -c Name)"
 
-echo "The following servers match the ${CAPO_TAG}:\n${SERVERS}"
-
-echo -e "\U0001F5D1 Pausing servers: ${SERVERS//$'\n'/ }"
-openstack ${OS_ARGS} server pause ${SERVERS//$'\n'/ }
-echo -e "\U0001F5D1 Deleting servers: ${SERVERS//$'\n'/ }"
-openstack ${OS_ARGS} server delete --wait ${SERVERS//$'\n'/ }
-echo -e "\U0001F5D1 Deleting volumes: ${SERVERS//$'\n'/-root }"
-openstack ${OS_ARGS} volume delete ${SERVERS//$'\n'/-root } --purge || true
+if [ -n "$SERVERS" ] ; then
+  echo -e "The following servers match the ${CAPO_TAG}:\n${SERVERS}\n"
+  
+  echo -e "\U0001F5D1 Pausing servers: ${SERVERS//$'\n'/ }"
+  openstack ${OS_ARGS} server pause ${SERVERS//$'\n'/ }
+  echo -e "\U0001F5D1 Deleting servers: ${SERVERS//$'\n'/ }"
+  openstack ${OS_ARGS} server delete --wait ${SERVERS//$'\n'/ }
+  echo -e "\U0001F5D1 Deleting volumes: ${SERVERS//$'\n'/-root }"
+  openstack ${OS_ARGS} volume delete ${SERVERS//$'\n'/-root } --purge || true
+else
+  echo -e "(no server to cleanup)"
+fi
 
 for i in $(seq 1 10)
 do
